:hammer: Collected predict methods
* All predict-methods have been moved to predict.rs
